Swap BaseClient and SafeClient methods
